user service split & service repositions
