+++ Chirp/src/app/shared/components/chirrup-list/chirrup-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { GetChirrupsService } from '../../services/get-chirrups.service';
+import { GetChirrupsService } from '../../../features/chirrup/services/get-chirrups.service';
 import { Chirrup, Comment, Content } from '../../../core/models/chirrup';
 import { CommentService } from 'src/app/features/chirrup/services/comment.service';
 import { SharedService } from 'src/app/features/chirrup/services/shared.service';
