+++ Chirp/src/app/shared/shared.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { ButtonModule } from 'primeng/button';
 import { CardModule } from 'primeng/card';
-import { GetChirrupsService } from '../shared/services/get-chirrups.service';
 import { FormsModule } from '@angular/forms';
 import { ReversePipe } from './pipes/reverse.pipe';
 import { Truncate1Pipe } from './pipes/truncate1.pipe';
